refactor: homogeneize API between all CPU backends.
